Fix some clang compiler warnings

+++ projects/amdsmi/rocm_smi/src/rocm_smi.cc
@@ -1436,7 +1436,7 @@ static rsmi_status_t get_frequencies(amd::smi::DevInfoTypes type, rsmi_clk_type_
   // "no current level" via f->current = -1 so callers can still report
   // the frequency table.
   if (f->current >= f->num_supported) {
-    f->current = -1;
+    f->current = static_cast<uint32_t>(-1);
   }
 
   return RSMI_STATUS_SUCCESS;
@@ -2252,7 +2252,7 @@ rsmi_status_t rsmi_dev_process_isolation_get(uint32_t dv_ind, uint32_t* pisolate
     LOG_ERROR(ss);
     return RSMI_STATUS_UNEXPECTED_DATA;
   }
-  *pisolate = partition_status[partition_id];
+  *pisolate = static_cast<uint32_t>(partition_status[partition_id]);
   return RSMI_STATUS_SUCCESS;
 }
 
@@ -2304,7 +2304,7 @@ rsmi_status_t rsmi_dev_process_isolation_set(uint32_t dv_ind, uint32_t pisolate)
   }
 
   // (3) Create the complete list with the update
-  partition_status[partition_id] = pisolate;
+  partition_status[partition_id] = static_cast<int>(pisolate);
   std::stringstream result;
   std::copy(partition_status.begin(), partition_status.end(),
             std::ostream_iterator<int>(result, " "));
@@ -2398,7 +2398,7 @@ rsmi_status_t rsmi_dev_xgmi_plpd_get(uint32_t dv_ind, rsmi_dpm_policy_t* policy)
       return RSMI_STATUS_UNEXPECTED_DATA;
     }
 
-    policy->policies[policy->num_supported].policy_id = value;
+    policy->policies[policy->num_supported].policy_id = static_cast<uint32_t>(value);
     std::string description = amd::smi::trim(tokens[1]);
     if (current_line.back() == '*') {  // current policy
       description.pop_back();          // remove last *
@@ -2503,7 +2503,7 @@ rsmi_status_t rsmi_dev_soc_pstate_get(uint32_t dv_ind, rsmi_dpm_policy_t* policy
       return RSMI_STATUS_UNEXPECTED_DATA;
     }
 
-    policy->policies[policy->num_supported].policy_id = value;
+    policy->policies[policy->num_supported].policy_id = static_cast<uint32_t>(value);
     std::string description = amd::smi::trim(tokens[1]);
     if (current_line.back() == '*') {  // current policy
       description.pop_back();          // remove last *
@@ -2642,7 +2642,7 @@ static rsmi_status_t get_dev_name_from_file(uint32_t dv_ind, char* name, size_t 
   rsmi_status_t ret = get_dev_value_line(amd::smi::kDevDevProdName, dv_ind, &val_str);
 
   if (ret != 0) {
-    return amd::smi::ErrnoToRsmiStatus(ret);
+    return amd::smi::ErrnoToRsmiStatus(static_cast<int>(ret));
   }
   size_t ct = val_str.copy(name, len);
 
@@ -3085,7 +3085,7 @@ rsmi_status_t rsmi_dev_pci_bandwidth_get(uint32_t dv_ind, rsmi_pcie_bandwidth_t*
       static_cast<uint32_t>(speed_index) * WIDTH_DATA_LENGTH + static_cast<uint32_t>(width_index);
   for (cur_index = 0; cur_index < WIDTH_DATA_LENGTH * SPEED_DATA_LENGTH; cur_index++) {
     b->transfer_rate.frequency[cur_index] =
-        static_cast<long>(link_speed[cur_index / WIDTH_DATA_LENGTH]) * 100 * 1000000L;
+        static_cast<uint64_t>(link_speed[cur_index / WIDTH_DATA_LENGTH]) * 100UL * 1000000UL;
     b->lanes[cur_index] = link_width[cur_index % WIDTH_DATA_LENGTH];
   }
   /*
